pyproject.toml
[project]
name = "kindle-notes"
version = "0.1.0"
description = "CLI tool to sync, browse, and search Kindle highlights"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.12"
keywords = ["kindle", "highlights", "notes", "ebook"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "binary-cookies-parser>=2.1.2",
    "click>=8.1",
    # kfxlib vendored deps (used by kfxlib/, not kn/ directly)
    "lxml>=6.0.2",
    "pillow>=12.1.1",
    "pypdf>=6.7.5",
    # browser scraping
    "playwright>=1.49",
]

[project.urls]
Repository = "https://github.com/delebedev/kindle-notes"

[project.scripts]
kn = "kn.cli:main"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["kn*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["kfxlib/"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
